requirements.txt
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
pymongo==4.6.1
python-dotenv==1.0.0